Correción de errores

+++ src/main/java/com/example/demo/service/serviceImpl/CursoServiceImpl.java
@@ -1,8 +1,6 @@
 package com.example.demo.service.serviceImpl;
 
-import com.example.demo.dao.CursoDAO;
-import com.example.demo.dao.EstadoDAO;
-import com.example.demo.dao.ProfesorDAO;
+import com.example.demo.dao.*;
 import com.example.demo.model.Curso;
 import com.example.demo.model.Estado;
 import com.example.demo.model.Profesor;
@@ -15,13 +13,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Service;
 
+import java.util.Date;
 import java.util.List;
 
 @Scope("singleton")
 @Service
 public class CursoServiceImpl implements CursoService {
 
-    //TODO: faltan validaciones y metodos del crud
     @Autowired
     CursoDAO cursoDAO;
 
@@ -32,23 +30,31 @@ public class CursoServiceImpl implements CursoService {
     ProfesorDAO profesorDAO;
 
     @Autowired
-    EstadoService estadoService;
+    RetoDAO retoDAO;
 
     @Autowired
-    ProfesorService profesorService;
+    MisionDAO misionDAO;
+
+    @Autowired
+    CursoEstudianteDAO cursoEstudianteDAO;
+
+
     @Override
     public String registrar(Curso curso) throws Exception {
-        validaciones(curso);
-        cursoDAO.save(curso);
-        return "Se creo exitosamente el curso.";
+        try {
+            validacionesCrear(curso);
+            cursoDAO.save(curso);
+            return "Se creo exitosamente el curso.";
+        }catch (Exception e){
+            return e.getMessage().toString();
+        }
+
     }
 
     @Override
     public String actualizar(CursoDTO cursoDTO) throws Exception {
         Curso curso = null;
-        Estado idEstado = null;
-        Profesor idProfesor = null;
-        //Todo: Faltan validaciones
+        validacionesActualizar(cursoDTO);
         curso = cursoDAO.findById(cursoDTO.getIdCurso()).orElse(null);
         curso.setNombre(cursoDTO.getNombre());
         curso.setPassword(cursoDTO.getPassword());
@@ -56,10 +62,10 @@ public class CursoServiceImpl implements CursoService {
         curso.setInicioCurso(cursoDTO.getInicioCurso());
         curso.setFinCurso(cursoDTO.getFinCurso());
         curso.setProgreso(cursoDTO.getProgreso());
-        idProfesor = profesorService.findById(cursoDTO.getIdProfesor());
-        curso.setProfesor(idProfesor);
-        idEstado = estadoService.findById(cursoDTO.getIdEstado());
-        curso.setEstado(idEstado);
+        curso.setProfesor(profesorDAO.findById(cursoDTO.getIdProfesor()).orElse(null));
+        curso.setEstado(estadoDAO.findById(cursoDTO.getIdEstado()).orElse(null));
+        curso.setFechaModificacion(cursoDTO.getFechaModificacion());
+        curso.setUsuarioModificador(cursoDTO.getUsuarioModificador());
         cursoDAO.save(curso);
         return "Se actualizo el curso.";
     }
@@ -72,7 +78,15 @@ public class CursoServiceImpl implements CursoService {
         if (!cursoDAO.existsById(idCurso)){
             throw new Exception("El curso con id: " + idCurso + " no existe.");
         }
-        //Todo: Faltan las validaciones de reto, mision y cursoEstudiante
+        if(!retoDAO.findByIdCurso(idCurso).isEmpty()){
+            throw new Exception("No se puede eliminar el curso porque esta siendo utilizado en un reto.");
+        }
+        if(!misionDAO.findByIdCurso(idCurso).isEmpty()){
+            throw new Exception("No se puede eliminar el curso porque esta siendo utilizado en una misión.");
+        }
+        if(!cursoEstudianteDAO.findByIdCurso(idCurso).isEmpty()){
+            throw new Exception("No se puede eliminar el curso porque esta siendo utilizado en un curso estudiante.");
+        }
         cursoDAO.deleteById(idCurso);
         return "Se elimino exitosamente el curso.";
     }
@@ -101,12 +115,135 @@ public class CursoServiceImpl implements CursoService {
         return cursoDAO.findById(idCurso).get();
     }
 
-    public void validaciones(Curso curso) throws Exception{
+    public void validacionesCrear(Curso curso) throws Exception{
+        if(curso.getNombre() == null || curso.getNombre().trim().equals("")){
+            throw new Exception("Debe ingresar el nombre del curso.");
+        }
+        if(Validaciones.isStringLenght(curso.getNombre(),50)){
+            throw new Exception("El nombre del curso es muy largo, solo se aceptan 50 caracteres.");
+        }
+        if(curso.getPassword() == null || curso.getPassword().trim().equals("")){
+            throw new Exception("Se debe ingresar una contraseña para el curso.");
+        }
+        if(Validaciones.isStringLenght(curso.getPassword(), 50)){
+            throw new Exception("La contraseña es muy larga, solo se aceptan 50 caracteres.");
+        }
+        if(curso.getCantidadEstudiantes()<1){
+            throw new Exception("Se debe ingresar la cantidad de estudiantes que tendrán acceso al curso.");
+        }
+        if(curso.getProgreso()!=0){
+            throw new Exception("Al crear el curso se crea con el progreso en el 0%.");
+        }
+        if(curso.getInicioCurso() == null){
+            throw new Exception("Debe ingresar una fecha de inicio del curso.");
+        }
+        Date fechaActual = new Date();
+        if(curso.getInicioCurso().compareTo(fechaActual)>0){
+            throw new Exception("Digite una fecha de inicio del curso válida.");
+        }
+        if(curso.getFinCurso().compareTo(fechaActual)>0){
+            throw new Exception("Digite una fecha de finalización del curso válida.");
+        }
+        if(curso.getFinCurso().compareTo(curso.getInicioCurso())>0){
+            throw new Exception("La fecha de finalización del curso no puede ser antes de la fecha de inicio.");
+        }
+        if(curso.getEstado().getIdEstado() == null){
+            throw new Exception("Debe ingresar un id estado.");
+        }
+        if(curso.getProfesor().getIdProfesor() == null){
+            throw new Exception("Debe ingresar un id del profesor.");
+        }
+        if(curso.getEstado().getIdEstado()<0 || curso.getEstado().getIdEstado().toString().equals("Optional.empty")){
+            throw new Exception("Debe ingresar un id estado válido.");
+        }
+        if(curso.getProfesor().getIdProfesor()< 0 || curso.getProfesor().getIdProfesor().toString().equals("Optional.empty")){
+            throw new Exception("Debe ingresar un id del profesor válido.");
+        }
         if(!estadoDAO.existsById(curso.getEstado().getIdEstado())){
             throw new Exception("El estado que ingreso no es válido.");
         }
         if(!profesorDAO.existsById(curso.getProfesor().getIdProfesor())){
             throw new Exception("El profesor que ingreso no es válido.");
         }
+        if(curso.getUsuarioCreador()==null || curso.getUsuarioCreador().equals("")){
+            throw new Exception("Debe ingresar el usuario creador.");
+        }
+        if(Validaciones.isStringLenght(curso.getUsuarioCreador(),50)){
+            throw new Exception("El nombre del usuario creador es muy largo, solo puede contener 50 caracteres.");
+        }
+        if(curso.getFechaCreacion()==null || curso.getFechaCreacion().toString().equals("")){
+            throw new Exception("Debe ingresar una fecha de creación.");
+        }
+    }
+
+    private void validacionesActualizar(CursoDTO cursoDTO)throws Exception{
+        if(cursoDTO.getIdCurso() == null || cursoDTO.getIdCurso().toString().equals("")){
+            throw new Exception("Se debe ingresar el id del curso que desea actualizar.");
+        }
+        if(!cursoDAO.existsById(cursoDTO.getIdCurso())){
+            throw new Exception("No se encontró el curso con ese id.");
+        }
+        if(!estadoDAO.existsById(cursoDTO.getIdEstado())){
+            throw new Exception("No se encontró el estado con ese id.");
+        }
+        if(!profesorDAO.existsById(cursoDTO.getIdProfesor())){
+            throw new Exception("No se encontró el profesor con ese id.");
+        }
+        if(cursoDTO.getNombre() == null || cursoDTO.getNombre().trim().equals("")){
+            throw new Exception("Debe ingresar el nombre del curso.");
+        }
+        if(Validaciones.isStringLenght(cursoDTO.getNombre(),50)){
+            throw new Exception("El nombre del curso es muy largo, solo se aceptan 50 caracteres.");
+        }
+        if(cursoDTO.getPassword() == null || cursoDTO.getPassword().trim().equals("")){
+            throw new Exception("Se debe ingresar una contraseña para el curso.");
+        }
+        if(Validaciones.isStringLenght(cursoDTO.getPassword(), 50)){
+            throw new Exception("La contraseña es muy larga, solo se aceptan 50 caracteres.");
+        }
+        if(cursoDTO.getCantidadEstudiantes()<1){
+            throw new Exception("Se debe ingresar la cantidad de estudiantes que tendrán acceso al curso.");
+        }
+        if(cursoDTO.getProgreso()!=0){
+            throw new Exception("Al crear el curso se crea con el progreso en el 0%.");
+        }
+        if(cursoDTO.getInicioCurso() == null){
+            throw new Exception("Debe ingresar una fecha de inicio del curso.");
+        }
+        Date fechaActual = new Date();
+        if(cursoDTO.getInicioCurso().compareTo(fechaActual)>0){
+            throw new Exception("Digite una fecha de inicio del curso válida.");
+        }
+        if(cursoDTO.getFinCurso().compareTo(fechaActual)>0){
+            throw new Exception("Digite una fecha de finalización del curso válida.");
+        }
+        if(cursoDTO.getFinCurso().compareTo(cursoDTO.getInicioCurso())>0){
+            throw new Exception("La fecha de finalización del curso no puede ser antes de la fecha de inicio.");
+        }
+        if(cursoDTO.getIdEstado() == null){
+            throw new Exception("Debe ingresar un id estado.");
+        }
+        if(cursoDTO.getIdProfesor() == null){
+            throw new Exception("Debe ingresar un id del profesor.");
+        }
+        if(cursoDTO.getIdEstado()<0 || cursoDTO.getIdEstado().toString().equals("Optional.empty")){
+            throw new Exception("Debe ingresar un id estado válido.");
+        }
+        if(cursoDTO.getIdProfesor()< 0 || cursoDTO.getIdProfesor().toString().equals("Optional.empty")){
+            throw new Exception("Debe ingresar un id del profesor válido.");
+        }
+        if(cursoDTO.getUsuarioModificador()==null || cursoDTO.getUsuarioModificador().equals("")){
+            throw new Exception("Debe ingresar el usuario modificador.");
+        }
+        if(Validaciones.isStringLenght(cursoDTO.getUsuarioModificador(),50)){
+            throw new Exception("El nombre del usuario modificador es muy largo, solo puede contener 50 caracteres.");
+        }
+        if(cursoDTO.getFechaModificacion()==null || cursoDTO.getFechaModificacion().toString().equals("")){
+            throw new Exception("Debe ingresar una fecha de modificación.");
+        }
+        if(cursoDTO.getFechaModificacion().compareTo(fechaActual)>0){
+            throw new Exception("No puede ingresar una fecha que aun no ha sucedido.");
+        }
+
     }
 }
